projects/cafe24-automation/requirements.txt
# Computer Use Agent Requirements
# Core dependencies for Claude Code Agent

# Computer Vision & Screen Capture
pillow>=10.0.0
opencv-python>=4.8.0
mss>=9.0.0
pytesseract>=0.3.10

# Automation & Control
pyautogui>=0.9.54
pynput>=1.7.6
keyboard>=0.13.5
mouse>=0.7.1

# AI & ML
anthropic>=0.34.0
openai>=1.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0

# Image Processing
scikit-image>=0.21.0
imageio>=2.31.0

# System & Utils
psutil>=5.9.0
pywin32>=306  # Windows specific
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.5.0

# Web & API
requests>=2.31.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0